chore: relocate & rename send screen components
